+++ packages/plugins/views/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @express-tool/plugin-views
 
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @express-tool/core@1.0.2
+
 ## 1.0.1
 
 ### Patch Changes